Fix "any" collision sensing
The comparator was performing an extra step. At best this was unnecessary; at worst it would stop detecting collisions.

The core comparison is `arcadephysics.touching & PROP`, where `PROP` is a constant such as `Kiwi.Components.Arcadephysics.LEFT`. The `touching` bitmask is of the form `0x1111` (that is the ANY form).

The `&` bitwise op will return the constant if the constant bit is in the mask. However, the ANY constant is multi-bit. It will return some positive number, but not necessarily ANY.

So if you check whether the `&` result equals the constant you passed, it will succeed for single-bit constants, but not ANY. This is quite unnecessary and breaks stuff. We just need to know whether the `&` was truthy or falsy.