Create plugin-style Goto URL block.
Our compiles aren't working. The Google closure-compiler service seems to be rejecting everything from this repo, including very old versions of the repo that built with no problems. It's possible that it is reacting to `//` formations in URLs as comment signifiers, but I'm unwilling to dig through all of an ancient Blockly branch to fix it.

Instead, let's use Blockly the way it was meant to be used: plugins!

Here's the Goto URL block again, built using the Blockly Developer Tools and code from the old version. It's now a little sterner about URL protocol: we officially frown upon links that don't begin with `http://` or `https://`.