Fix and deprecate kiwi_get_editor_sounds block.
Because it derives its dropdown menu from the current game's sound population, and because the menu is stored by value, AND because the sound population changes on remix, a remixed game with this block will break badly. It will fail to compile for preview, and you cannot open the faulty script to fix it.

These fixes allow you to preview again, although you still cannot open the script. We must hope that people will not abuse the block.