Fix "camera shake" not returning to 0.
The problem was, it didn't _start_ from 0. It started from whatever the offset was, and then altered the offset. Overlapping shakes would thus accumulate wandering origins, and faithfully return to the wrong place. I've fixed this by the rough measure of forcing the origin to be 0, discarding any current camera offset. I also rounded the output, which mutes the later camera movements slightly, but also avoids tilemaps and textures having edge freak.